Switch to a looser eval strategy for rendering tag bodies.
The previous method relied on a stack of attributes Hashes in the RenderState;
the new method in effect wraps a Binding with a method_missing for translating
identifiers that look like locals back into template attributes. This not only
allows for template attributes to be used as arguments to the methods called
on other attributes, but avoids sharing the Binding of the RenderState with
the tags rendered inside it.

Backwards-compatibility is also maintained via a few aliases and shim methods.